+++ src/components/DragDrop/DragDropTypes.ts
@@ -21,4 +21,5 @@ export interface ContainerType {
   itemIds: string[];
   header?: React.ReactNode;
   emptyContent?: React.ReactNode;
+  columnClassName?: string[];
 }